requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.4.4
propcache==0.3.2
PyYAML==6.0.2
yarl==1.20.1
